// spacetime-presence-ts/example/src/module_bindings/app/auth/types.ts
export const AuthUser = __t.object("AuthUser", {
  userId: __t.string(),
  email: __t.string(),
  emailVerified: __t.bool(),
  name: __t.option(__t.string()),
  image: __t.option(__t.string()),
  createdAt: __t.timestamp(),
  updatedAt: __t.timestamp(),
});
export type AuthUser = __Infer<typeof AuthUser>;